+++ src/layouts/Main/components/Sidebar/Sidebar.js
@@ -3,7 +3,7 @@ import clsx from 'clsx';
 import PropTypes from 'prop-types';
 import { makeStyles } from '@material-ui/styles';
 import { Divider, Drawer } from '@material-ui/core';
-import DashboardIcon from '@material-ui/icons/Dashboard';
+import ShoppingCartIcon from '@material-ui/icons/ShoppingCart';
 import LockOpenIcon from '@material-ui/icons/LockOpen';
 import KitchenIcon from '@material-ui/icons/Kitchen';
 
@@ -39,9 +39,9 @@ const Sidebar = props => {
 
   const pages = [
     {
-      title: 'Dashboard',
-      href: '/dashboard',
-      icon: <DashboardIcon />
+      title: 'Shopping List',
+      href: '/shopping-list',
+      icon: <ShoppingCartIcon />
     },
     {
       title: 'Item',