+++ src/views/Items/components/ItemTable/ItemTable.js
@@ -13,7 +13,7 @@ import {ItemRow} from './components'
 
 const useStyles = makeStyles(theme => ({
   hiddenColumns: {
-    [theme.breakpoints.down('xs')]: {
+    [theme.breakpoints.down('md')]: {
       display: 'none'
     }
   },
@@ -26,14 +26,12 @@ const ItemTable = ({categories}) => {
       <Table aria-label="collapsible table">
         <TableHead>
           <TableRow>
-            <TableCell/>
-            <TableCell>Name</TableCell>
-            <TableCell align="right">
-              Items</TableCell>
-            <TableCell align="right">Shopping List</TableCell>
-            <TableCell align="right" className={classes.hiddenColumns}>Last
-              Update</TableCell>
-            <TableCell></TableCell>
+            <TableCell style={{maxWidth:'20px'}}></TableCell>
+            <TableCell style={{maxWidth:'20px'}}>Name</TableCell>
+            <TableCell align="right" className={classes.hiddenColumns}>Items</TableCell>
+            <TableCell align="right" className={classes.hiddenColumns}>Shopping List</TableCell>
+            <TableCell align="right" className={classes.hiddenColumns}>Last Update</TableCell>
+            <TableCell style={{maxWidth:'20px'}}></TableCell>
           </TableRow>
         </TableHead>
         <TableBody>