+++ src/views/ShoppingList/components/index.js
@@ -1,0 +1,4 @@
+export { default as ShoppingListCard } from './ShoppingListCard';
+export { default as ShoppingListToolbar } from './ShoppingListToolbar';
+export { default as CreateEditShoppingListDialog } from './CreateEditShoppingListDialog';
+export { default as DeleteDialog } from './DeleteDialog';