+++ src/views/index.js
@@ -1,6 +1,6 @@
 export { default as Maintenance } from './Maintenance';
 export { default as NotFound } from './NotFound';
 export { default as SignIn } from './SignIn';
-export { default as Dashboard } from './Dashboard';
+export { default as ShoppingList } from './ShoppingList';
 export { default as SignOut } from './SignOut';
 export { default as Items } from './Items';